# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reading Quest API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /worlds/{worldId}/stories:
    get:
      operationId: listStoriesInWorld
      tags: [worlds]
      parameters:
        - in: path
          name: worldId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Stories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Story" }
        "404":
          description: World not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /stories/{storyId}:
    get:
      operationId: getStory
      tags: [worlds]
      parameters:
        - in: path
          name: storyId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Story with chapter list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoryDetail" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /chapters/{chapterId}:
    get:
      operationId: getChapter
      tags: [worlds]
      parameters:
        - in: path
          name: chapterId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Chapter content
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chapter" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Story:
      type: object
      properties:
        id: { type: integer }
        worldId: { type: integer }
        slug: { type: string }
        title: { type: string }
        summary: { type: string }
        chapterCount: { type: integer }
        finishedChapterCount: { type: integer }
      required: [id, worldId, slug, title, summary, chapterCount, finishedChapterCount]
    StoryDetail:
      type: object
      properties:
        id: { type: integer }
        worldId: { type: integer }
        title: { type: string }
        summary: { type: string }
        chapters:
          type: array
          items: { $ref: "#/components/schemas/ChapterSummary" }
      required: [id, worldId, title, summary, chapters]
    ChapterSummary:
      type: object
      properties:
        id: { type: integer }
        storyId: { type: integer }
        sortIndex: { type: integer }
        title: { type: string }
        finished: { type: boolean }
      required: [id, storyId, sortIndex, title, finished]
    Chapter:
      type: object
      properties:
        id: { type: integer }
        storyId: { type: integer }
        sortIndex: { type: integer }
        title: { type: string }
        sceneImageUrl: { type: string }
        chapterCount: { type: integer }
        nextChapterId:
          type: ["integer", "null"]
        paragraphs:
          type: array
          items: { type: string }
        tappableWords:
          type: array
          items: { $ref: "#/components/schemas/TappableWord" }
      required: [id, storyId, sortIndex, title, sceneImageUrl, chapterCount, paragraphs, tappableWords]
    TappableWord:
      type: object
      properties:
        key: { type: string }
        word: { type: string }
        syllables:
          type: array
          items: { type: string }
        reaction: { type: string }
      required: [key, word, syllables, reaction]
...