# lib/api-zod/src/generated/types/chapter.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Quest API
 * OpenAPI spec version: 0.1.0
 */
import type { TappableWord } from "./tappableWord";

export interface Chapter {
  id: number;
  storyId: number;
  sortIndex: number;
  title: string;
  sceneImageUrl: string;
  chapterCount: number;
  /** @nullable */
  nextChapterId?: number | null;
  paragraphs: string[];
  tappableWords: TappableWord[];
}