# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reading Quest API
 * OpenAPI spec version: 0.1.0
 */

export interface Story {
  id: number;
  worldId: number;
  slug: string;
  title: string;
  summary: string;
  chapterCount: number;
  finishedChapterCount: number;
}